+++ src/pages/admin/staff/Staff.jsx
@@ -2,11 +2,12 @@ import { Link } from "react-router-dom";
 import { useDispatch, useSelector } from "react-redux";
 import React, { useEffect, useState } from "react";
 import { deleteStaff, getAllStaff } from "../../../services/staff.service";
+import Loader from "../../../components/Loader";
 
 const Student = () => {
   const dispatch = useDispatch();
 
-  const { staffs } = useSelector((state) => state.staff);
+  const { staffs, loading } = useSelector((state) => state.staff);
 
   const [searchTerm, setSearchTerm] = useState("");
 
@@ -15,7 +16,9 @@ const Student = () => {
   );
 
   function handleDelete(id) {
-    dispatch(deleteStaff(id));
+    if (window.confirm("Are you sure you want to delete?")) {
+      dispatch(deleteStaff(id));
+    }
   }
 
   useEffect(() => {
@@ -24,6 +27,7 @@ const Student = () => {
 
   return (
     <section className="p-3 sm:p-4 rounded-lg w-full h-auto mt-[10px] sm:px-8">
+      {loading && <Loader />}
       <div className="flex flex-col sm:flex-row sm:items-center gap-4 sm:gap-0 justify-between px-2 py-4">
         <div className="flex justify-between sm:justify-center items-center gap-4 text-xl sm:text-[1.4rem] font-semibold">
           <div>
@@ -39,7 +43,7 @@ const Student = () => {
 
         <input
           type="search"
-          placeholder="Search Student by name"
+          placeholder="Search staff by name"
           className="border border-gray-400 rounded-md p-2 text-sm"
           value={searchTerm}
           onChange={(e) => setSearchTerm(e.target.value)}
@@ -49,17 +53,23 @@ const Student = () => {
       <div id="overflow" className="overflow-x-auto ">
         <table className="min-w-full text-left table-auto border-collapse text-[0.83rem] whitespace-nowrap">
           <thead>
-            <tr className="bg-gray-700 text-gray-100 text-primary">
-              {["SR#", "Name", "Email", "Phone no", "Att %", "Role", "Actions"].map(
-                (header) => (
-                  <th
-                    key={header}
-                    className="text-[0.92rem] py-3 px-4 border-b border-secondary"
-                  >
-                    {header}
-                  </th>
-                )
-              )}
+            <tr className="bg-[#212121] text-gray-100 text-primary">
+              {[
+                "SR#",
+                "Name",
+                "Email",
+                "Phone no",
+                "Att %",
+                "Role",
+                "Actions",
+              ].map((header) => (
+                <th
+                  key={header}
+                  className="text-[0.92rem] py-3 px-4 border-b border-secondary"
+                >
+                  {header}
+                </th>
+              ))}
             </tr>
           </thead>
 
@@ -114,7 +124,7 @@ const Student = () => {
 
         {filteredStaffs.length === 0 && (
           <div className="h-[50vh] flex justify-center items-center text-sm">
-            No Student Found
+            No Staff Found
           </div>
         )}
       </div>
